+++ src/pymizer/__init__.py
@@ -8,7 +8,7 @@ from ._bridge import (
     runtime_diagnostics,
 )
 from .datasets import NorthSeaExample, list_datasets, load_dataset, load_north_sea
-from .model import MizerParams, MizerSim, new_community_params, new_multispecies_params, new_single_species_params, read_params
+from .model import MizerParams, MizerSim, new_community_params, new_multispecies_params, new_single_species_params, read_params, read_rds
 from ._validation import validate_interaction_matrix, validate_species_params
 
 __all__ = [
@@ -26,6 +26,7 @@ __all__ = [
     "new_multispecies_params",
     "new_single_species_params",
     "read_params",
+    "read_rds",
     "runtime_diagnostics",
     "validate_interaction_matrix",
     "validate_species_params",